+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 7.14.3 (unreleased)
 ## Implementation changes and bug fixes
 - [PR #1218](https://github.com/rqlite/rqlite/pull/1218): Check for more possible errors in peers.json. Thanks @Tjstretchalot
+- [PR #1220](https://github.com/rqlite/rqlite/pull/1220): Support Notify over Raft connection.
 
 ## 7.14.2 (April 7th 2023)
 This release is the first to includes various bug fixes and optimizations thanks to running much of the code through [Chat GPT-4](https://openai.com/product/gpt-4), most of which are not explicitly listed in the [CHANGELOG](https://github.com/rqlite/rqlite/edit/master/CHANGELOG.md), but you can check the commit history for details. Future releases of rqlite will probably include more such changes.